requirements.txt
Django~=3.1.7
python-socketio==5.1.0
numpy==1.20.1